pyproject.toml
[project]
name = "ebirders-pt"
version = "2025.06.28"
description = "Data and statistics for checklists submitted to eBird Portugal"
readme = "README.md"
license-files = ["LICENSE.txt"]
requires-python = ">=3.12"
dependencies = [
    "django-environ>=0.11.2",
    "django-extensions>=3.2.3",
    "django-spaceless-templates>=1.0.2",
    "django>=5.1.3",
    "gunicorn>=23.0.0",
    "ipython>=8.29.0",
    "psycopg2-binary>=2.9.10",
    "pymemcache>=4.0.0",
    "python-dateutil>=2.9.0.post0",
    "whitenoise[brotli]>=6.8.2",
    "ebird-codes>=0.0.2",
    "ebird-api>=3.4.2",
    "requests>=2.32.3",
    "lxml>=5.3.2",
    "beautifulsoup4>=4.13.4",
    "sentry-sdk>=2.26.1",
    "django-autocomplete-light>=3.12.1",
    "django-ckeditor>=6.7.3",
    "ebird-api-data>=0.1.0",
]

[tool.ruff]
fix = true
src = [".", "apps"]
extend-exclude = [
    "**/migrations/*.py"
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
lines-between-types = 1
section-order = [
    "future",
    "standard-library",
    "django",
    "third-party",
    "first-party",
    "local-folder",
    "testing",
]

[tool.ruff.lint.isort.sections]
"django" = ["django"]
"testing" = ["faker", "pytest"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
FAIL_INVALID_TEMPLATE_VARS = true
testpaths = [
    "apps"
]
log_cli = true

[tool.bumpversion]
current_version = "2025.06.28"
parse = """(?x)                        # Verbose mode
    (?P<release>                       # The release part
        (?:[1-9][0-9]{3})\\.           # YYYY.
        (?:1[0-2]|0[1-9])\\.           # 0M.
        (?:3[0-1]|[1-2][0-9]|0[1-9])   # 0D
    )
    (?:\\.(?P<patch>\\d+))?            # .patch, optional
"""
serialize = ["{release}.{patch}", "{release}"]
ignore_missing_version = false
tag = true
sign_tags = true
tag_name = "v{new_version}"
tag_message = "New version: {current_version} → {new_version}"
allow_dirty = false
commit = true
commit_args = "--no-verify"
message = "New version: {current_version} → {new_version}"
pre_commit_hooks = ["uv sync --upgrade", "git add uv.lock"]

[tool.bumpversion.parts.release]
calver_format = "{YYYY}.{0M}.{0D}"

[[tool.bumpversion.files]]
filename = "config/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "CHANGELOG.md"
search = """
## Latest
"""
replace = """
## Latest

## {new_version}
"""

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[tool.uv]
dev-dependencies = [
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
    "pytest-django>=4.9.0",
    "django-debug-toolbar>=4.4.6",
    "faker>=37.1.0",
    "ruff>=0.11.5",
    "bump-my-version>=1.1.2",
    "matplotlib>=3.10.1",
    "jupyterlab>=4.4.1",
    "factory-boy>=3.3.3",
]
